Add XRC01 exemple
